added context menu for notification icon | split prev. awklib.h to window.h and notification.h
